\name{CircE-package}
\alias{CircE}
\docType{package}
\title{Circumplex models Estimation}
\description{
  An R package for circumplex models estimation.  This package contains functions for fitting circumplex structural models for correlation matrices (with negative correlation) by the method of maximum likelihood.}

\details{
\tabular{ll}{
Package: \tab CircE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014/3/4 \cr
License: \tab  GPL (>= 2)\cr
}}

\author{
Michele Grassi \email{grassi.mic@gmail.com}

}




