\name{Verbal}
\alias{Verbal}
\title{ Verbal Ability Tests correlation matrix }

\description{Intercorrelations of six Verbal Ability Tests for 1046 Bucknell College Sophomores.}
\usage{data(Verbal)}
\source{
Guttman, L.(1954). A new approach to factor analysis. the radex. In P. F. Lazasrsfeld (Ed.), \emph{Mathematical thinking in the social sciences} (pp. 258-348). New York:Columbia University Press.\cr
Browne, M. W. (1992) Circumplex models for correlation matrices. \emph{Psychometrika}, 57, pp. 469-497.

}
\examples{
data(Verbal)
Verbal

}